requirements.txt
asgiref==3.3.1
Django==3.0
django-cors-headers==3.6.0
django-mssql-backend==2.8.1
djangorestframework==3.12.2
mysqlclient==2.0.3
psycopg2-binary==2.8.6
pyodbc==4.0.30
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.3
